putting board into gameComponent as its Constraints are independent from Cell
