--- src/main/scala/de/htwg/se/ludo/model/boardComponent/boardBaseImpl/board.md
@@ -1,21 +1,0 @@
-#Sample board representation:
-
-- Y: Yellow
-- G: Green
-- R: Red
-- B: Black
-
-##Base:
-<!-- Base position
-0       3  4      7 8      11  12    15 -->
-[Y][][][] [G][][][] [R][][][] [B][][][]
-
-##Game:
-<!-- Start positions
-16                     26                      36                      46                  55 -->
-[Y][][][] [][][][] [][][G][] [][][][] [][][][] [R][][][] [][][][] [][][B][] [][][][] [][][][]
-
-##Home:
-<!-- Home positions
-56        60        64        68     71 --> 
-[Y][][][] [G][][][] [R][][][] [B][][][]
