+++ src/main/scala/de/htwg/se/ludo/model/gameComponent/gameBaseImpl/BasicBoardConstraints.scala
@@ -1,0 +1,6 @@
+package de.htwg.se.ludo.model.gameComponent.gameBaseImpl
+
+object BasicBoardConstraints {
+  val fields = 72
+  val filling: Cell = EmptyCell
+}
