+++ src/main/scala/de/htwg/se/ludo/model/gameComponent/gameBaseImpl/Cell.scala
@@ -1,4 +1,6 @@
-package de.htwg.se.ludo.model.gameComponent
+package de.htwg.se.ludo.model.gameComponent.gameBaseImpl
+
+import de.htwg.se.ludo.model.gameComponent.CellInterface
 
 sealed case class Cell(value: String) extends CellInterface {
   def isSet:Boolean = value != ""
